add project structure/dependencies finish scripts
I had to move a few things around and add the copy webpack plugin in
order to get icons and images to copy nicely into the dist folder. I
also have effectively completed most if not all of the scripting I am
going to need to do for this project. Only thing left is to start
styling things.